Update: Fix compiler warning
git-svn-id: https://svn.code.sf.net/p/moneymanagerex/code/trunk@2537 7f5d68d0-c712-0410-839f-9b116a576cc5

+++ mmex/src/transdialog.cpp
@@ -347,7 +347,7 @@ void mmTransDialog::CreateControls()
         wxTRANSLATE("Transfer")
     };
     // Restrict choise if accounts number less than 2
-    int size = sizeof(transaction_type)/sizeof(wxString);
+    size_t size = sizeof(transaction_type)/sizeof(wxString);
     if (core_->getNumBankAccounts() < 2) size--;
     for(size_t i = 0; i < size; ++i)
     choiceTrans_->Append(wxGetTranslation(transaction_type[i]),
